01/01/2022 22:55

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: alukongo <alukongo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/22 15:36:07 by alukongo          #+#    #+#             */
-/*   Updated: 2021/12/30 15:36:14 by alukongo         ###   ########.fr       */
+/*   Updated: 2022/01/01 19:59:23 by alukongo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,6 +17,8 @@
 # include<stdlib.h>
 # include<stdio.h>
 # include<stdarg.h>
+#include<limits.h>
+#include<stdint.h>
 
 int		ft_strlen(char *s);
 int		ft_putchar(char c);
@@ -27,7 +29,7 @@ int		ft_unsigned(unsigned int nb);
 char	*ft_uitoa(unsigned int nb);
 int		nb_digit(unsigned int nb, int base);
 int		ft_printf(const char *str, ...);
-int		print_address(int nbr, char *base);
-void	ft_putnbr_hex(int nbr, char *base);
+int		print_address(unsigned long long nbr, char *base);
+void	ft_putnbr_hex(unsigned long long nbr, char *base);
 
 #endif
